355-rename-lagom-dev-mode-components: move all classes in the devmode package and also move back the LagomDevModeComponents in the devmode package
