+++ dev/service-registry/devmode-scaladsl/src/main/scala/com/lightbend/lagom/devmode/internal/scaladsl/registry/ScalaServiceRegistryClient.scala
@@ -2,11 +2,10 @@
  * Copyright (C) Lightbend Inc. <https://www.lightbend.com>
  */
 
-package com.lightbend.lagom.internal.scaladsl.registry
+package com.lightbend.lagom.devmode.internal.scaladsl.registry
 
 import java.net.URI
 
-import com.lightbend.lagom.devmode.internal.scaladsl.registry.ServiceRegistry
 import com.lightbend.lagom.devmode.internal.registry.AbstractLoggingServiceRegistryClient
 import com.lightbend.lagom.scaladsl.api.transport.NotFound
